Tests/LibWeb/Text/expected/canvas/webgl-draw-invalidates-cached-command.txt

[webgl] cached command has new generation: true
[webgl2] cached command has new generation: true
